[server] ファイル分割
